requirements.txt
streamlit==1.49.1
pillow==11.3.0
python-dotenv==1.1.1
azure-cognitiveservices-vision-computervision==0.9.1
msrest==0.7.1
azure-common==1.1.28
azure-core>=1.24.0
pandas>=1.4.0
pyarrow>=7.0